backend/requirements.txt
yfinance
pandas
python-dotenv
pinecone-client
langchain
langchain_google_genai
